+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
   * Add wc middleware (render web components)
   * Add routerLinks middleware
+  * Rename `next` middleware hook to `resolve`
   * Export Router as named export instead of as default
   * Convert classes to ES6 syntax
   * Do not transpile to ES5 (publishes package as ES6)